feat: getAlerts717Performance to filter by diseaseName in data layer
