+++ src/domain/usecases/Get717PerformanceUseCase.ts
@@ -5,7 +5,6 @@ import {
     PerformanceMetrics717,
     PerformanceMetrics717Key,
 } from "../entities/disease-outbreak-event/PerformanceOverviewMetrics";
-import { Future } from "../entities/generic/Future";
 import { Id } from "../entities/Ref";
 import { PerformanceOverviewRepository } from "../repositories/PerformanceOverviewRepository";
 
@@ -28,40 +27,7 @@ export class Get717PerformanceUseCase {
         } else if (type === "national") {
             return this.options.performanceOverviewRepository.getNational717Performance();
         } else if (type === "alerts") {
-            return this.options.performanceOverviewRepository
-                .getAlerts717Performance()
-                .flatMap(performanceMetrics =>
-                    this.getAlerts717PerformanceMetrics(diseaseName, performanceMetrics)
-                );
+            return this.options.performanceOverviewRepository.getAlerts717Performance(diseaseName);
         } else throw new Error(`Unknown 717 type: ${type} `);
     }
-
-    private getAlerts717PerformanceMetrics(
-        diseaseName: Maybe<DiseaseNames>,
-        performanceMetrics: PerformanceMetrics717[]
-    ): FutureData<PerformanceMetrics717[]> {
-        if (!diseaseName) return Future.success(performanceMetrics);
-
-        const secondaryDiseaseMetrics = performanceMetrics.filter(
-            metric => metric.type === "secondary" && metric.disease === diseaseName
-        );
-
-        const primaryDiseaseMetrics = secondaryDiseaseMetrics.map<PerformanceMetrics717>(
-            metric => ({
-                ...metric,
-                type: "primary",
-                value: calculatePrimaryDiseaseValueFromSecondaryValue(metric),
-            })
-        );
-
-        return Future.success([...primaryDiseaseMetrics, ...secondaryDiseaseMetrics]);
-    }
-}
-
-function calculatePrimaryDiseaseValueFromSecondaryValue(
-    metric: PerformanceMetrics717
-): number | "Inc" {
-    return metric.value !== undefined && metric?.total !== undefined && metric.value !== "Inc"
-        ? parseFloat((metric.value / metric.total).toFixed(2))
-        : "Inc";
 }
